requirements.txt
# Core scientific stack
numpy>=1.24,<3
pandas>=2.0,<3
matplotlib>=3.7,<4

# NLP and evaluation
nltk>=3.8,<4
sacrebleu>=2.4,<3
rouge-score>=0.1.2,<0.2
evaluate>=0.4,<0.5
python-Levenshtein>=0.23,<0.26

# Hugging Face / modeling
transformers>=4.40,<5
accelerate>=0.27,<0.30
bitsandbytes>=0.43,<0.44
# Note: Install a CUDA-appropriate torch build if needed
torch>=2.1,<3

# OpenAI / Azure OpenAI client
openai>=1.30,<2

# Utilities
requests>=2.31,<3
tqdm>=4.66,<5
tabulate>=0.9,<1
jsonlines>=4.0,<5
python-dotenv>=1.0,<2

# Optional but commonly used for notebooks
jupyter>=1.0,<2
ipykernel>=6.29,<7
